Add SFTP static IP and IP filtering
This commit uses a VPC endpoint for the SFTP transfer service, instead
of the PUBLIC option, and configures a VPC with a NLB and static IP. It
exposes the DNS name of the NLB so that other packages can create CNAME
records to it. It also adds a security group for ingress IP filtering.

*IMPORTANT*: I have not fully tested the IP filtering functionality, and
I have a suspicion that it doesn't work for the allow all: 0.0.0.0/0.