switch dlm snapshot role name to arn
This swaps out the optional role name for an optional role arn and removes the
data resource.  This fixes issues with dependencies since a role name is treated
as a non-computed value and the data resource can be evaluated too early.  The
local default is used to allow a blank default when the variable is passed
through multiple layers.