luthername tags output
This adds a tags output to allow directly passing the tags into a resource
without having to enumerate them.  In the case where replication is greater than
one, the name and ID will not be contained in the tags and the tags output can
be merged with the indexed name and ID values.

This also adjusts the behavior of the optional inputs to result in null outputs
if the input isn't supplied to avoid unexpected behavior from accidentally using
blank strings.