fix for previous luthername change
A previous change to luthername had unprovided tags result in null outputs
instead of an empty string.  That change broke tagging that explicitly
referenced an optional tag output with a null value.  This fixes the affected
ASG resource by using dynamic tag blocks based on the luthername `tags` output.
The consequences of this are:
- Name tag propagates to the instances
- ID and Resource tags will be added

The luthername `tags` map output is also fixed to exclude null values to prevent
unexpected consequences when iterating over it.